i#1680 large pages: fix yet more Windows breakage from 326a0d5
Fixes Windows test build breakage from 326a0d5.

Fixes Windows test execution breakage from 326a0d5: aslr-ind turned out to
have a reliance on the init sequence that 326a0d5 broke.

Review-URL: https://codereview.appspot.com/311860043

+++ suite/tests/security-common/TestMemProtChg.c
@@ -1,5 +1,5 @@
 /* **********************************************************
- * Copyright (c) 2013-2015 Google, Inc.  All rights reserved.
+ * Copyright (c) 2013-2016 Google, Inc.  All rights reserved.
  * Copyright (c) 2003-2004 VMware, Inc.  All rights reserved.
  * **********************************************************/
 
@@ -94,7 +94,7 @@ test_alloc_overlap(void)
      * modifying mmap can have its prot match the region base's prot to
      * make it harder for DR to detect.
      */
-    char *buf = allocate_mem(PAGE_SIZE*2, ALLOW_READ|ALLOW_WRITE|ALLOW_EXEC);
+    char *buf = allocate_mem((int)PAGE_SIZE*2, ALLOW_READ|ALLOW_WRITE|ALLOW_EXEC);
     char *code = copy_to_buf(buf + PAGE_SIZE, PAGE_SIZE, NULL, CODE_INC, COPY_NORMAL);
     protect_mem(code, PAGE_SIZE, ALLOW_READ|ALLOW_EXEC);
     test_print(code, 42);
